+++ NEWS.md
@@ -7,6 +7,7 @@
 * `e3-convert-xunit`: add testcase-level `<system-out>` and
   `<system-err>` handling.
 * `e3-convert-xunit`: robustify for unexpected status tags.
+* `e3-convert-xunit`: copy capped messages to result logs.
 
 27.2 (2025-06-17)
 =================